Add back button from settings, use partial
Tried using the ~H syntax and defining components in the view, but that
didn't seem to work too well for this use case (iterating over a list of
members, rendering a component for each item in the list).

Instead, went with a partial for displaying each person it's shared with.